Ajout des coords parents dans la liste des individus rattachés
